Basic react integration for project creator
Successfully added basic react functionality to project creation. If more is added to the .jsx file, use 'npm run webpack' in a terminal to create a usable build in the /lib folder (will also need to use 'npm install' in the terminal to install the dependencies needed to do so).