Progress on displaying role selectors with react
Current progress towards using react to display role selectors. Only issue present is that number input values do not update correctly when changed (It will properly update the value in the roleList array, but the value inside the element remains the same). Wanted to upload since we may be switching focuses away from prototypes for now.